+++ src/settings.ts
@@ -10,10 +10,10 @@ export function getMethodOpeningBraceOnNewLine(): string {
     return '';
 }
 
-export function isIncludeFluentWithSetters(): boolean | undefined {
+export function isIncludeFluentWithSetters(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('includeFluentWithSettersGetters') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('includeFluentWithSettersGetters')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('includeFluentWithSettersGetters')! as boolean)
     );
 }
 
@@ -21,44 +21,51 @@ export function getFluentMethodPrefix(): string {
     return vscode.workspace.getConfiguration('java.code.generators').get('fluentMethodPrefix') || '';
 }
 
-export function isGenerateEvenIfExists(): boolean | undefined {
+export function isGenerateEvenIfExists(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('generateEvenIfExists') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('generateEvenIfExists')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('generateEvenIfExists')! as boolean)
     );
 }
 
-export function isOnlyIdForHashAndEquals(): boolean | undefined {
+export function isOnlyIdForHashAndEquals(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('onlyIdForHashAndEquals') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('onlyIdForHashAndEquals')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('onlyIdForHashAndEquals')! as boolean)
     );
 }
 
-export function isOnlyPrimitiveForToString(): boolean | undefined {
+export function isOnlyPrimitiveForToString(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('onlyPrimitiveForToString') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('onlyPrimitiveForToString')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('onlyPrimitiveForToString')! as boolean)
     );
 }
 
-export function copyJsonPropertyAnnotationsFromVariablesToSettersGetters(): boolean | undefined {
+export function copyJsonPropertyAnnotationsFromVariablesToSettersGetters(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('copyJsonPropertyAnnotationsFromVariablesToSettersGetters') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('copyJsonPropertyAnnotationsFromVariablesToSettersGetters')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('copyJsonPropertyAnnotationsFromVariablesToSettersGetters')! as boolean)
     );
 }
 
-export function includeGeneratedAnnotation(): boolean | undefined {
+export function includeGeneratedAnnotation(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('includeGeneratedAnnotation') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('includeGeneratedAnnotation')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('includeGeneratedAnnotation')! as boolean)
     );
 }
 
-export function fluentCallsNormalSetters(): boolean | undefined {
+export function fluentCallsNormalSetters(): boolean {
     return (
         vscode.workspace.getConfiguration('java.code.generators').has('fluentCallsNormalSetters') &&
-        vscode.workspace.getConfiguration('java.code.generators').get('fluentCallsNormalSetters')
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('fluentCallsNormalSetters')! as boolean)
+    );
+}
+
+export function includeBeta(): boolean {
+    return (
+        vscode.workspace.getConfiguration('java.code.generators').has('includeBeta') &&
+        (vscode.workspace.getConfiguration('java.code.generators').get<boolean>('includeBeta')! as boolean)
     );
 }